fixing the update issue bug and working on uploading group image api

+++ PaaS/src/main.rs
@@ -205,7 +205,7 @@ async fn main() -> MainResult<(), Box<dyn std::error::Error + Send + Sync + 'sta
         access_level_cli = &args[2];
     }
     if username_cli != &"".to_string() && access_level_cli != &"".to_string(){
-        match utils::set_user_access(username_cli.to_owned(), access_level_cli.parse::<u8>().unwrap(), db.clone()).await{
+        match utils::set_user_access(username_cli.to_owned(), access_level_cli.parse::<i64>().unwrap(), db.clone()).await{
             Ok(user_info) => {
                 info!("access level for user {} has been updated successfully", username_cli);
                 info!("updated user {:?}", user_info);
